requirements-backend.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# MongoDB
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Data Science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
shap==0.44.0

# Utilities
python-dotenv==1.0.0
